+++ CHANGELOG.md
@@ -40,6 +40,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Added `GET /token` endpoint for SAML authorization when cumulus is protected by Launchpad.
     This lets a user retieve a token by hand that can be presented to the API.
 
+- **CUMULUS-1513**
+  - Added `sqs`-type rule support in the Cumulus API `@cumulus/api`
+  - Added `sqsMessageConsumer` lambda which processes messages from the SQS queues configured in the `sqs` rules.
+
 ### Changed
 
 - **CUMULUS-1453**